improve feedback
In normal operation, the messages are sent without triggering notifications.
The message for the opened window is only sent once it is actually open,
the receipt of the message is only acknowledged by sending a typing marker.

Normally, a users should be standing in front of the fablock and should thus
be able to tell that it is operating, independent of feedback on Telegram.