add TLS to networking
Just having TCP and relying on the connection being stable is not
a good security model. Instead mutual TLS is used to authenticate
both ends during communication. It also encrypts the traffic which
is not really necessary, but this also provides authentication.

Since the certificates might be hard to set up and the networking
might not be desired, also added a configuration variable to more
easily disable all networking.